rpk: move logdirs command from debug to cluster
