+++ src/go/rpk/pkg/cli/cmd/cluster/logdirs.go
@@ -7,7 +7,7 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0
 
-package debug
+package cluster
 
 import (
 	"context"
@@ -26,7 +26,7 @@ import (
 	"github.com/twmb/types"
 )
 
-func newLogdirsCommand(fs afero.Fs) *cobra.Command {
+func NewLogdirsCommand(fs afero.Fs) *cobra.Command {
 	var (
 		configFile string
