+++ src/go/rpk/pkg/cli/cmd/debug/debug.go
@@ -23,7 +23,6 @@ func NewCommand(fs afero.Fs) *cobra.Command {
 	cmd.AddCommand(
 		newBundleCommand(fs),
 		NewInfoCommand(fs),
-		newLogdirsCommand(fs),
 	)
 
 	return cmd
